requirements.txt
keyboard
pyautogui
pytesseract
requests
opencv-python
numpy
Pillow
win10toast